Move `InMemoryActorProvider` to a test package equivalent to `ActorProvider`'s
