+++ core-test/src/main/java/br/com/orcinus/orca/core/test/auth/actor/InMemoryActorProvider.kt
@@ -13,7 +13,7 @@
  * not, see https://www.gnu.org/licenses.
  */
 
-package br.com.orcinus.orca.core.test
+package br.com.orcinus.orca.core.test.auth.actor
 
 import br.com.orcinus.orca.core.auth.actor.Actor
 import br.com.orcinus.orca.core.auth.actor.ActorProvider
