+++ core/src/test/java/br/com/orcinus/orca/core/auth/actor/ActorProviderTests.kt
@@ -15,8 +15,8 @@
 
 package br.com.orcinus.orca.core.auth.actor
 
-import br.com.orcinus.orca.core.test.InMemoryActorProvider
 import br.com.orcinus.orca.core.test.auth.Authenticator
+import br.com.orcinus.orca.core.test.auth.actor.InMemoryActorProvider
 import kotlin.test.assertEquals
 import kotlinx.coroutines.test.runTest
 import org.junit.Test
